refactor(brokers): Rename redirect.js to oauth-redirect.js
This name is a bit more explicit.
