+++ app/scripts/models/auth_brokers/oauth-redirect.js
